refactored around price/proof add pages
New file:
* `price_add_helper.dart`: Helper around price/proof adding pages.

Impacted files
* `product_price_add_page.dart`: refactored with new class `PriceAddHelper`
* `proof_bulk_add_page.dart`: refactored with new class `PriceAddHelper`
